+++ common/src/main/java/de/the_build_craft/remote_player_waypoints_for_xaero/common/waypoints/MutablePlayerPosition.java
@@ -31,13 +31,13 @@ import java.util.UUID;
 public class MutablePlayerPosition {
     private final String id;
     public final UUID uuid;
-    public final MutableFloat3 pos;
+    public final MutableDouble3 pos;
     private WaypointState waypointState;
 
     public MutablePlayerPosition(PlayerPosition playerPosition, WaypointState waypointState) {
         this.id = playerPosition.id;
         this.uuid = playerPosition.gameProfile.getId();
-        this.pos = new MutableFloat3(playerPosition.pos);
+        this.pos = new MutableDouble3(playerPosition.pos);
         this.waypointState = waypointState;
     }
